v0.16.0 (#98)
* bump to v0.16.0

* Convert properties to js (#68)

* Convert properties to js

This involves rewiring things a bit. lnd now listens for cleartext http inside the container. It uses ports 8081 and 10010 so it doesn't conflict with proxyboi, which has been added as a reverse proxy for https conversion outside the container. lnd.embassy:8080 and 10009 are still used to connect from outside the container and https is also still used for outside the container. lncli apparently does not have the ability to connect via cleartext http, and it also doesn't seem to like the eOS system-generated cert, which we are now using, so lncli no longer works. Easiest fix for this is to add localhost to the system cert.

This also reworks the build using .dockerignore, so builds take less than a minute on my mac m1.

* address pr comments

* fix fresh installs and properties build

* don't copy entire wrapper repo into dockerfile

* fix properites for fresh nodes, fix x86, revert accidental change of bitcoin core default to proxy

* refactor properties.ts

---------

Co-authored-by: islandbitcoin <dread@start9.com>
Co-authored-by: Dread <34528298+islandbitcoin@users.noreply.github.com>

* set alpine ver to 3.16

* formatting and ssl cert

* use system cert

* remove LAN interface from manifest

* update submodule

* add rpcmiddleware.enable to lnd.conf

* Update manifest.yaml

Co-authored-by: Lucy C <12953208+elvece@users.noreply.github.com>

* clean up main.rs

---------

Co-authored-by: islandbitcoin <dread@start9.com>
Co-authored-by: Chris Guida <chrisguida@users.noreply.github.com>
Co-authored-by: Aiden McClelland <me@drbonez.dev>
Co-authored-by: Lucy C <12953208+elvece@users.noreply.github.com>